+++ app/build.gradle
@@ -70,6 +70,10 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
+    //material
+    implementation 'com.google.android.material:material:1.2.0-alpha1'
+//    implementation'com.makeramen:roundedimageview:2.3.0'
+
     //google
     implementation 'com.google.android.gms:play-services-auth:19.0.0'
 //    implementation 'com.google.android.gms:play-services-drive:11.6.0'
