+++ app/src/main/java/com/example/recorder/MainActivity.java
@@ -19,10 +19,10 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void run() {
                 // This method will be executed once the timer is over
-                Intent i = new Intent(MainActivity.this, Home.class);
+                Intent i = new Intent(MainActivity.this, RecordsHome.class);
                 startActivity(i);
                 finish();
             }
-        }, 2000);
+        },3000);
     }
 }
