+++ app/src/main/java/com/example/recorder/drop/UploadTask.java
@@ -15,9 +15,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-
 import static com.example.recorder.storage.Constant.Call_Records;
-import static com.example.recorder.storage.Constant.Drop_Box_Date;
 
 public class UploadTask extends AsyncTask{
