+++ app/src/main/res/drawable/button_sheet_background.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <solid android:color="@color/white"/>
+
+    <corners android:topLeftRadius="28dp"
+        android:topRightRadius="28dp"/>
+
+</shape>
