+++ app/src/main/res/drawable/layout_selector.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <selector xmlns:android="http://schemas.android.com/apk/res/android">
+        <item android:state_focused="true"
+            android:drawable="@color/theme"/>
+        <item android:state_pressed="true"
+            android:state_enabled="false"
+            android:drawable="@color/card_color" />
+        <item android:drawable="@color/white"/>
+    </selector>
+
+</selector>
