+++ app/src/main/res/drawable/shape.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <corners
+        android:topLeftRadius="8dp"
+        android:topRightRadius="8dp"
+        android:bottomRightRadius="8dp"
+        android:bottomLeftRadius="8dp"/>
+
+    <stroke
+        android:width="1dp"
+        android:color="@color/theme"/>
+
+    <solid android:color="@color/theme"/>
+
+</shape>
