--- app/src/main/res/layout/activity_home.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/drawer_layout"
-    android:fitsSystemWindows="true"
-    tools:openDrawer="start"
-    tools:context=".Home">
-
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:theme="?actionBarTheme"
-                android:minHeight="?attr/actionBarSize"
-                app:navigationIcon="@drawable/menu"/>
-
-
-
-            <androidx.appcompat.widget.SwitchCompat
-                android:id="@+id/switchButton"
-                android:layout_width="wrap_content"
-                android:layout_centerInParent="true"
-                android:layout_height="wrap_content"
-                android:checked="false"
-                android:textAppearance="?android:attr/textAppearanceMedium"
-                android:text="Recording"
-                />
-
-
-        </RelativeLayout>
-
-
-    <com.google.android.material.navigation.NavigationView
-        android:id="@+id/nav_view"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        app:headerLayout="@layout/header"
-        app:menu="@menu/drawer_view"
-        android:fitsSystemWindows="true"
-        android:layout_gravity="start"/>
-
-
-
-
-</androidx.drawerlayout.widget.DrawerLayout>
