--- app/src/main/res/layout/activity_one_drive.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".onedrive.OneDrive">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/one_drive_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="?attr/actionBarTheme"
-        app:navigationIcon="@drawable/back"
-        tools:ignore="MissingConstraints" >
-
-    </androidx.appcompat.widget.Toolbar>
-
-    <Button
-        android:id="@+id/one_drive_signIn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="SigIn"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/one_drive_toolbar" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
