--- app/src/main/res/layout/activity_setting.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".Setting">
-
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar_setting"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="?attr/actionBarTheme"
-        app:navigationIcon="@drawable/back"/>
-
-
-    <LinearLayout
-        android:id="@+id/linear_setting"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:padding="10dp"
-        android:visibility="visible"
-        android:layout_below="@+id/toolbar_setting">
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Please select your storage"
-            android:textColor="@color/black"
-            android:textSize="22sp"
-            android:textStyle="bold" />
-
-        <RadioGroup
-            android:id="@+id/radio_group"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="5dp"
-            android:text="Radio"
-            android:textSize="18sp">
-
-            <RadioButton
-                android:id="@+id/google"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Google drive"
-                android:textSize="18sp" />
-
-            <RadioButton
-                android:id="@+id/drop_box"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="DropBox"
-                android:textSize="18sp" />
-
-            <RadioButton
-                android:id="@+id/oneDrive"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="One Drive"
-                android:textSize="18sp" />
-
-            <RadioButton
-                android:id="@+id/local"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Local Storage"
-                android:textSize="18sp" />
-
-        </RadioGroup>
-
-        <Button
-            android:id="@+id/submit"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15dp"
-            android:text="click ok"
-            android:background="#000000"
-            android:textSize="20sp" />
-
-    </LinearLayout>
-
-
-
-</LinearLayout>
