--- app/src/main/res/layout/header.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="20dp">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Header"
-        android:layout_marginTop="10dp"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        android:textColor="@color/black"/>
-
-</LinearLayout>
