--- app/src/main/res/menu/drawer_view.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <group android:checkableBehavior="single">
-
-        <item
-            android:id="@+id/nav_first_fragment"
-            android:icon="@drawable/home"
-            android:title="Home" />
-
-        <item
-            android:id="@+id/nav_second_fragment"
-            android:icon="@drawable/setting"
-            android:title="Setting" />
-
-
-    </group>
-
-</menu>
