+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,11 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+
+    <color name="orange">#F47238</color>
+<!--    <color name="header">#131b20</color>-->
+    <color name="header">#10289B</color>
+        <color name="theme">#2946D3</color>
+<!--    <color name="card_color">#27363f</color>-->
+    <color name="card_color">#10289B</color>
 </resources>
